Made cloud shader tile better
I can only offset on the x axis, not both. Otherise the pipe edges don't match. The whole thing is still very repeating, because of all the tiling I'm doing.
I can introduce a different type of noise rotating in the opposite direction  but must find a good way to blend them together.