Collistions work
Placed a rigi body on the player root object so now triggers are registered and the paradox vanishes or does "damage".
Fiddled with the shaders a bit more.
Need to divide the rotation offset passed to the shader by the tiling on that axis. Maybe I can do this in the shader.