Implemented basic HUD
The hud displays a progress bar with moments. When the bar reachoes zero it's game over. Picking up moment refils part of the bar. Hitting paradox reduces it.
Score is represented as a date. The more distance you make the more you go back in time. Picking up moments icreases the multiplyer (how fast the date goes back in time).